+++ packages/react/src/types/syn-jsx-elements.ts
@@ -295,12 +295,15 @@ export type SynCustomElement<
   ]
 >;
 /**
- * @summary Comboboxes allow you to choose items from a menu of predefined options.
+ * @summary A combobox component that combines the functionality of a text input with a dropdown listbox,
+ * allowing users to either select from predefined options or enter custom values (when not restricted).
+ *
  * @documentation https://synergy-design-system.github.io/?path=/docs/components-syn-combobox--docs
  * @status stable
  *
  * @dependency syn-icon
  * @dependency syn-popup
+ * @dependency syn-tag
  *
  * @slot - The listbox options. Must be `<syn-option>` elements.
  *    You can use `<syn-optgroup>`'s to group items visually.
@@ -330,7 +333,7 @@ export type SynCustomElement<
  * @csspart form-control-label - The label's wrapper.
  * @csspart form-control-input - The combobox's wrapper.
  * @csspart form-control-help-text - The help text's wrapper.
- * @csspart combobox - The container the wraps the prefix, combobox, clear icon, and expand button.
+ * @csspart combobox - The container that wraps the prefix, combobox, clear icon, and expand button.
  * @csspart prefix - The container that wraps the prefix slot.
  * @csspart suffix - The container that wraps the suffix slot.
  * @csspart display-input - The element that displays the selected option's label,
@@ -343,6 +346,12 @@ export type SynCustomElement<
  * @csspart popup - The popup's exported `popup` part.
  * Use this to target the tooltip's popup container.
  * @csspart no-results - The container that wraps the "no results" message.
+ * @csspart tags - The container that houses option tags when `multiple` is used.
+ * @csspart tag - The individual tags that represent each selected option in `multiple`.
+ * @csspart tag__base - The tag's base part.
+ * @csspart tag__content - The tag's content part.
+ * @csspart tag__remove-button - The tag's remove button.
+ * @csspart tag__remove-button__base - The tag's remove button base part.
  *
  * @animation combobox.show - The animation to use when showing the combobox.
  * @animation combobox.hide - The animation to use when hiding the combobox.
@@ -1672,12 +1681,15 @@ declare module 'react' {
        * @csspart form-control-help-text - The help text's wrapper.
        */ 'syn-checkbox': SynCheckboxJSXElement;
       /**
-       * @summary Comboboxes allow you to choose items from a menu of predefined options.
+       * @summary A combobox component that combines the functionality of a text input with a dropdown listbox,
+       * allowing users to either select from predefined options or enter custom values (when not restricted).
+       *
        * @documentation https://synergy-design-system.github.io/?path=/docs/components-syn-combobox--docs
        * @status stable
        *
        * @dependency syn-icon
        * @dependency syn-popup
+       * @dependency syn-tag
        *
        * @slot - The listbox options. Must be `<syn-option>` elements.
        *    You can use `<syn-optgroup>`'s to group items visually.
@@ -1707,7 +1719,7 @@ declare module 'react' {
        * @csspart form-control-label - The label's wrapper.
        * @csspart form-control-input - The combobox's wrapper.
        * @csspart form-control-help-text - The help text's wrapper.
-       * @csspart combobox - The container the wraps the prefix, combobox, clear icon, and expand button.
+       * @csspart combobox - The container that wraps the prefix, combobox, clear icon, and expand button.
        * @csspart prefix - The container that wraps the prefix slot.
        * @csspart suffix - The container that wraps the suffix slot.
        * @csspart display-input - The element that displays the selected option's label,
@@ -1720,6 +1732,12 @@ declare module 'react' {
        * @csspart popup - The popup's exported `popup` part.
        * Use this to target the tooltip's popup container.
        * @csspart no-results - The container that wraps the "no results" message.
+       * @csspart tags - The container that houses option tags when `multiple` is used.
+       * @csspart tag - The individual tags that represent each selected option in `multiple`.
+       * @csspart tag__base - The tag's base part.
+       * @csspart tag__content - The tag's content part.
+       * @csspart tag__remove-button - The tag's remove button.
+       * @csspart tag__remove-button__base - The tag's remove button base part.
        *
        * @animation combobox.show - The animation to use when showing the combobox.
        * @animation combobox.hide - The animation to use when hiding the combobox.